requirements.txt
numpy
opencv-python
scikit-learn
tensorflow
python-dotenv
Pillow
colorama